// src/app/countries/interface/cache-storage.interface.ts
import { Country } from './contry.interface';
import { Region } from './region.type';

export interface CacheStoreInterface {
  byCapital: TermCountries;
  byCountries: TermCountries;
  byRegion: RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: Country[];
}

export interface RegionCountries {
  region: Region;
  countries: Country[];
}